+++ packages/orq-rc/src/models/operations/updateidentity.ts
@@ -167,7 +167,7 @@ export const UpdateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-17T15:38:15.072Z",
+    "2026-02-17T16:30:56.053Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {